hashCode , equals

+++ practice03/src/main/java/prob06/Div.java
@@ -8,7 +8,7 @@ public class Div extends Arith{
 	
 
 	@Override
-	public int calculate() {
-		return getA()/getB();
+	public Double calculate() {
+		return (double) getA()/getB();
 	}
 }
